+++ frontend/src/assets/documents/changelog.md
@@ -1,3 +1,10 @@
+## v2.2.1
+_May 31st, 2024_
+
+- Added the ability to paste a Bestiary Builder statblock directly into the current statblock if it was exported using the export button.
+  - This means you no longer need to open a menu and paste the json directly. Simply press export on one page, and press paste on the other.
+- Fixes an issue where Animal Handling and Sleight of Hand skills would not show on the statblock since v2.0.0.
+
 ## v2.2.0
 _May 29th, 2024_